==> 2-gen_traces.ipynb <==
# Generate Traces

Code authored by: Shaw Talebi

### imports

import pandas as pd
from dotenv import load_dotenv
load_dotenv()

from utils.tools import *
from utils.gen_data import send_openai_request, generate_system_prompt, send_tai_messages, num_tools_available
from utils.tool_calling import parse_tool_call, call_tool, format_tool_result

import re
import json

### load data

df = pd.read_csv('data/queries.csv')

# Convert query_type to categorical with custom order
df['query_type'] = pd.Categorical(df['query_type'], categories=['no_tool', 'easy', 'hard'], ordered=True)
df = df.sort_values('query_type')

### initialize data structures

trace_list = []
num_tools_list = []

### Type 1: No Tool Call

df_notool = df[df['query_type'] == 'no_tool']

%%time
for row in df_notool.iterrows():
    print(f"Processing row {row[0]}")
    
    # create system prompt
    num_tools_input = random.choice(['single', 'few', 'many'])
    num_tools = num_tools_available(num_tools_input)
    system_prompt = generate_system_prompt(None, num_tools)

    # grab query
    query = row[1]['query']

    # construct message_list
    message_list = [
        {"role": "user", "content": query},
    ]
    
    # generate response
    response = send_tai_messages(message_list)

    # add system message and response to message_list
    message_list.insert(0, {"role": "user", "content": system_prompt})
    message_list.append({"role": "assistant", "content": response})

    # update data structures
    trace_list.append(message_list)
    num_tools_list.append(num_tools)

print(len(trace_list))
print(len(num_tools_list))

### Type 2: Easy Tool Call

df_easy = df[df['query_type'] == 'easy']
tool_call_list = []

#### generate tool call

%%time
for row in df_easy.iterrows():    
    print(f"Processing row {row[0]}")

    # generate system message
    tool_name = row[1]['tool_name']
    system_prompt = generate_system_prompt(tool_name, 0)
    query = row[1]['query']

    # generate tool call using GPT-4.1
    response = send_openai_request(system_prompt, query, temperature=0.5)[0]
    tool_name, tool_args = parse_tool_call(response)    

    # add tool call to list
    tool_call_list.append(response)

#### generate trace

%%time
for query, response in zip(df_easy['query'].to_list(), tool_call_list):

    # create system prompt
    num_tools_input = random.choice(['none', 'few', 'many'])
    num_tools = num_tools_available(num_tools_input)
    system_prompt = generate_system_prompt(tool_name, num_tools)

    # call tool
    tool_name, tool_args = parse_tool_call(response)
    result = call_tool(tool_name, tool_args)
    tool_result = format_tool_result(result)

    # construct message_list with tool_call
    message_list = [
        {"role": "user", "content": system_prompt},
        {"role": "user", "content": query},
        {"role": "assistant", "content": response},
        {"role": "user", "content": tool_result},
    ]

    final_response = send_tai_messages(message_list)

    # add final response to message_list
    message_list.append({"role": "assistant", "content": final_response})

    # update data structures
    trace_list.append(message_list)
    num_tools_list.append(num_tools)

print(len(trace_list))
print(len(num_tools_list))

### Type 3: Hard Tool Call

df_hard = df[df['query_type'] == 'hard']
think_tool_call_list = []

#### generate tool call (with some thinking)

%%time
for row in df_hard.iterrows():    
    print(f"Processing row {row[0]}")

    # generate system message
    tool_name = row[1]['tool_name']
    system_prompt = generate_system_prompt(tool_name, 0, filepath='prompts/system_cot.md')
    query = row[1]['query']

    # generate tool call using GPT-4.1
    response = send_openai_request(system_prompt, query, temperature=0.5)[0]
    tool_name, tool_args = parse_tool_call(response)

    # add tool call to list
    think_tool_call_list.append(response)

#### generate trace

%%time
for query, response in zip(df_hard['query'].to_list(), think_tool_call_list):

    # create system prompt
    num_tools_input = random.choice(['none', 'few', 'many'])
    num_tools = num_tools_available(num_tools_input)
    system_prompt = generate_system_prompt(tool_name, num_tools)

    # parse tool call
    tool_name, tool_args = parse_tool_call(response)
    try:
        result = call_tool(tool_name, tool_args)
        tool_result = format_tool_result(result)
    except:
        print("Could not get tool result for: ", tool_name, tool_args)
        tool_result = "Error: Could not get tool result."

    # construct message_list with tool_call
    message_list = [
        {"role": "user", "content": system_prompt},
        {"role": "user", "content": query},
        {"role": "assistant", "content": response},
        {"role": "user", "content": tool_result},
    ]

    final_response = send_tai_messages(message_list)

    # add final response to message_list
    message_list.append({"role": "assistant", "content": final_response})

    # update data structures
    trace_list.append(message_list)
    num_tools_list.append(num_tools)

### write traces to file

# add trace and num_tools to dataframe
df.insert(2, "trace", trace_list[:600])
df.insert(3, "num_tools_available", num_tools_list)

# write to .csv
df.to_csv('data/traces.csv', index=False)
