==> 3-data_prep.ipynb <==
# Data Preparation

Code authored by: Shaw Talebi

### imports

import pandas as pd
from utils.data_prep import add_answer_tags, view_trace, replace_trace_messages
from datasets import DatasetDict, Dataset

from dotenv import load_dotenv
load_dotenv()

### load data

df = pd.read_csv('data/traces.csv')

df.head()

### fix ad hoc issues

#### add final answer tags

# update final responses
for idx, trace in enumerate(df['trace']):
    df.at[idx, 'trace'] = add_answer_tags(trace, idx)

#### Fix rows: 344, 484, 486, 595

# row 344
new_final_response = """<final_answer>
75 divided by 8 22.0
</final_answer>"""

replace_trace_messages(df, 344, [{'index': 4, 'role': 'assistant', 'content': new_final_response}])

# Row 484
new_final_response = """<final_answer>
The Unix timestamp for 2002-11-25 is 1038204000.
</final_answer>"""

replace_trace_messages(df, 344, [{'index': 4, 'role': 'assistant', 'content': new_final_response}])

# Row 486
new_response = """You want to convert the date "2023-17-12" to a Unix timestamp. However, that is not a valid date.
I will assume you meant December 17, 2023 and use the get_unix_timestamp tool to perform this conversion.
<tool_call>
{
  "tool_name": "get_unix_timestamp",
  "args": {
    "date": "2023-12-17"
  }
}
</tool_call>"""

new_tool_result = """<tool_result>
1702792800
</tool_result>"""

new_final_response = """<final_answer>
The Unix timestamp for 2023-12-17 is 1702792800.
</final_answer>"""

replace_trace_messages(df, 486, [{'index': 2, 'role': 'assistant', 'content': new_response}, 
                                {'index': 3, 'role': 'user', 'content': new_tool_result},
                                {'index': 4, 'role': 'assistant', 'content': new_final_response},])

# Row 595
new_final_response = """<final_answer>
The slug for "New Product Launch!!! Exciting Times@2024" is "new-product-launch-exciting-2024".
</final_answer>"""
replace_trace_messages(df, 595, [{'index': 4, 'role': 'assistant', 'content': new_final_response},])

#### drop rows: 322, 424, 438

print(len(df))
df = df.drop(index=[322, 424, 438])
print(len(df))

### train-val-test split

# Shuffle the dataset
df_shuffled = df.sample(frac=1, random_state=42).reset_index(drop=True)

# Split into train, validation, and test sets (e.g., 80% train, 10% validation, 10% test)
train_frac = 0.8
valid_frac = 0.1
test_frac = 0.1

# define train and validation size
n_samples = len(df_shuffled)
train_size = int(train_frac * n_samples)
valid_size = int(valid_frac * n_samples) + 1

# create train, validation, and test datasets
df_train = df_shuffled[:train_size]
df_valid = df_shuffled[train_size:train_size + valid_size]
df_test = df_shuffled[train_size + valid_size:]

print(f"Training set: {len(df_train)} samples ({len(df_train)/n_samples:.1%})")
print(f"Validation set: {len(df_valid)} samples ({len(df_valid)/n_samples:.1%})")
print(f"Test set: {len(df_test)} samples ({len(df_test)/n_samples:.1%})")

### upload to HF hub

# Convert the pandas DataFrames back to Hugging Face Datasets
train_ds = Dataset.from_pandas(df_train)
valid_ds = Dataset.from_pandas(df_valid)
test_ds = Dataset.from_pandas(df_test)

# Combine into a DatasetDict
dataset_dict = DatasetDict({
    'train': train_ds,
    'validation': valid_ds,
    'test': test_ds
})

# push data to hub
dataset_dict.push_to_hub("shawhin/tool-use-finetuning")
